chore: Removed unnecessary slash commands
removed /ban
removed /faq
combined /status and /clearstatus
renamed twitchBan to streamBan
moved /membercount to /info role
removed /info channel
removed message info context menu
