+++ MomentumDiscordBot/Commands/Moderator/ModeratorModule.cs
@@ -9,7 +9,6 @@ using DSharpPlus.Entities;
 using MomentumDiscordBot.Constants;
 using MomentumDiscordBot.Models;
 using MomentumDiscordBot.Services;
-using MomentumDiscordBot.Commands.Autocomplete;
 
 namespace MomentumDiscordBot.Commands.Moderator
 {
@@ -25,15 +24,6 @@ namespace MomentumDiscordBot.Commands.Moderator
             StreamMonitorService.UpdateCurrentStreamersAsync(null);
 
             await ReplyNewEmbedAsync(context, "Updating Livestreams", MomentumColor.Blue);
-        }
-
-        [SlashCommand("ban", "Bans a user, purging their messages")]
-        public static async Task BanAsync(InteractionContext context, [Option("member", "member")] DiscordUser user)
-        {
-            DiscordMember member = (DiscordMember)user;
-            await member.BanAsync(7, $"Banned by {context.User} using !ban");
-            await ReplyNewEmbedAsync(context, $"Banned {member}, purging their messages in the last 7 days.",
-                MomentumColor.Red);
         }
 
         [SlashCommand("bans", "Returns a list of banned users")]
@@ -78,40 +68,5 @@ namespace MomentumDiscordBot.Commands.Moderator
                 .AddFile(fileName, fileStream));
         }
 
-        [SlashCommand("membercount", "Get the number of members with a role")]
-        public static async Task GetMembersWithRoleAsync(InteractionContext context, [Option("role", "role")] DiscordRole role)
-        {
-            var (_, guildRole) = context.Guild.Roles.FirstOrDefault(x => x.Key == role.Id);
-
-            if (guildRole != null)
-            {
-                var membersWithRole = context.Guild.Members.Values.Count(x => x.Roles.Contains(guildRole));
-                await ReplyNewEmbedAsync(context, $"{membersWithRole} users have {guildRole.Mention}",
-                    MomentumColor.Blue);
-            }
-            else
-            {
-                await ReplyNewEmbedAsync(context, "That role does not exist in this server", DiscordColor.Orange);
-            }
-        }
-
-        [SlashCommand("status", "Sets the bots status")]
-        public static async Task StatusAsync(InteractionContext context,
-            [Option("status", "status")] string status,
-            [ChoiceProvider(typeof(ActivityTypeChoiceProvider))][Option("type", "ActivityType")] string type = null)
-        {
-            var activity = Enum.TryParse(type, out ActivityType activityType)
-                ? new DiscordActivity(status, activityType)
-                : new DiscordActivity(status);
-            await context.Client.UpdateStatusAsync(activity);
-            await ReplyNewEmbedAsync(context, $"Status set to '{status}'.", MomentumColor.Blue);
-        }
-
-        [SlashCommand("clearstatus", "Clears the bots status")]
-        public static async Task ClearStatusAsync(InteractionContext context)
-        {
-            await context.Client.UpdateStatusAsync(new DiscordActivity());
-            await ReplyNewEmbedAsync(context, "Status cleared.", MomentumColor.Blue);
-        }
     }
 }
